+++ src/object/Information.cpp
@@ -68,8 +68,8 @@ void Information::setValue(const InfoValue val) {
     setValueImpl(val);
   } catch (const std::bad_variant_access &e) {
     throw std::invalid_argument(
-        "Invalid quality, please provide an instance of the matching "
-        "information value\n - Previous: " +
+        "Invalid value, please provide an instance of the matching "
+        "information object\n - Previous: " +
         InfoValue_toString(getValue()) +
         "\n - Proposed: " + InfoValue_toString(val));
   }
@@ -91,7 +91,7 @@ void Information::setQuality(const InfoQuality val) {
   } catch (const std::bad_variant_access &e) {
     throw std::invalid_argument(
         "Invalid quality, please provide an instance of the matching "
-        "information value\n - Previous: " +
+        "information object\n - Previous: " +
         InfoQuality_toString(getQuality()) +
         "\n - Proposed: " + InfoQuality_toString(val));
   }